extract init_common from skeleton/init and system/init
moved to init_common
  - init_tc
  - init_usb
  - init_i2c_master
  - init_i2c_slave (skeleton version)
