+++ earthsea/config.mk
@@ -66,6 +66,7 @@ CSRCS = \
        ../skeleton/adc.c     \
        ../skeleton/events.c     \
        ../skeleton/init.c \
+       ../skeleton/init_common.c \
        ../skeleton/i2c.c \
        ../skeleton/interrupts.c \
        ../skeleton/monome.c \
