+++ earthsea/main.c
@@ -20,6 +20,7 @@
 #include "types.h"
 #include "events.h"
 #include "init.h"
+#include "init_common.h"
 #include "interrupts.h"
 #include "monome.h"
 #include "timers.h"
