+++ meadowphysics/main.c
@@ -17,6 +17,7 @@
 #include "events.h"
 #include "i2c.h"
 #include "init.h"
+#include "init_common.h"
 #include "interrupts.h"
 #include "monome.h"
 #include "timers.h"
