+++ system/init.h
@@ -3,9 +3,5 @@
 
 extern void init_gpio(void);
 extern void init_spi(void);
-extern void init_tc(void);
-extern void init_usb_host (void);
-extern void init_i2c_master(void);
-extern void init_i2c_slave(void);
 
 #endif
