+++ teletype/main.c
@@ -25,6 +25,7 @@
 #include "types.h"
 #include "events.h"
 #include "init.h"
+#include "init_common.h"
 #include "interrupts.h"
 #include "i2c.h"
 #include "kbd.h"
