+++ whitewhale/main.c
@@ -23,6 +23,7 @@
 #include "events.h"
 #include "i2c.h"
 #include "init.h"
+#include "init_common.h"
 #include "interrupts.h"
 #include "monome.h"
 #include "timers.h"
